Two more

+++ dali/operators/image/color/brightness_contrast_test.cu
@@ -92,7 +92,7 @@ void BrightnessContrastVerify(TensorListWrapper input, TensorListWrapper output,
   float out_range = FullRange<OutputType>();
 
   ASSERT_EQ(input_tl->num_samples(), output_tl->num_samples());
-  for (size_t t = 0; t < input.cpu().num_samples(); t++) {
+  for (int t = 0; t < input.cpu().num_samples(); t++) {
     auto out_shape = output_tl->tensor_shape(t);
     auto out_tensor = output_tl->tensor<OutputType>(t);
     auto in_shape = input_tl->tensor_shape(t);
