+++ dali/operators/ssd/box_encoder.cu
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2021, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright (c) 2017-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -252,7 +252,7 @@ BoxEncoder<GPUBackend>::CalculateDims(
   TensorListShape<> boxes_output_shape(boxes_input.num_samples(), kBoxesOutputDim);
   TensorListShape<> labels_output_shape(boxes_input.num_samples(), kLabelsOutputDim);
 
-  for (size_t i = 0; i < boxes_input.num_samples(); i++) {
+  for (int i = 0; i < boxes_input.num_samples(); i++) {
     boxes_output_shape.set_tensor_shape(i,
         {anchor_count_, static_cast<int64_t>(BoundingBox::size)});
     labels_output_shape.set_tensor_shape(i, {anchor_count_});
